add user-is-not-authorized test
